Add files via upload
Fixed snake self-collision testing to maintain a ttl-dictionary instead of relying on a field array. The last card in the snake snaps into place when it moves in the same direction as the previous card; this is still weird occasionally, but it doesn't crash the game anymore.

Insertion sorting needs to be re-coded for new algorithms.